1.8v/main_test.go
package main

import (
	"os"
	"os/exec"
	"os/signal"
	"syscall"
	"testing"
	"time"
)

func TestMain(m *testing.M) {
	// Run the main function in a goroutine
	go func() {
		main()
	}()

	// Wait for a second to allow the main goroutine to start
	time.Sleep(time.Second)

	// Run the tests
	exitCode := m.Run()

	// Send a termination signal to the main goroutine
	pid := os.Getpid()
	process, _ := os.FindProcess(pid)
	process.Signal(syscall.SIGTERM)

	// Wait for the main goroutine to exit
	time.Sleep(time.Second)

	// Exit with the same exit code as the tests
	os.Exit(exitCode)
}

func TestMainFunction(t *testing.T) {
	// Set up a SIGTERM channel to receive termination signal
	signalCh := make(chan os.Signal, 1)
	signal.Notify(signalCh, syscall.SIGTERM)

	// Send a SIGTERM signal to terminate the main goroutine
	pid := os.Getpid()
	process, _ := os.FindProcess(pid)
	process.Signal(syscall.SIGTERM)

	// Wait for the termination signal
	select {
	case <-signalCh:
		// The main goroutine should exit gracefully
	case <-time.After(5 * time.Second):
		t.Fatal("Main goroutine did not exit gracefully")
	}
}

func TestRestartProgram(t *testing.T) {
	// Get the path to the test executable
	executablePath, err := os.Executable()
	if err != nil {
		t.Fatalf("Failed to get executable path: %v", err)
	}

	// Launch a new instance of the program with the same arguments
	cmd := exec.Command(executablePath, os.Args[1:]...)
	cmd.Stdout = os.Stdout
	cmd.Stderr = os.Stderr

	// Start the new instance of the program
	err = cmd.Start()
	if err != nil {
		t.Fatalf("Failed to start a new instance of the program: %v", err)
	}

	// Wait for the new instance to exit
	err = cmd.Wait()
	if err != nil {
		t.Fatalf("New instance of the program exited with an error: %v", err)
	}
}

// Add more tests as needed for the remaining functions
